Working Fund, with probably bugs
Included the investment fund. Consumers savings are invested in a fund
that finance capital purchases and the occasional loss of firms. In
return consumers get a share of the profits. Non invested savings enjoy
an exogenous rent.
So far the model fails to takeoff. Probably some bug in the capital
investment routines, since firms fail to invest in capital.